change array method from map to filter
Based on the variable name "completedTodos", the variable value should be an array of **todo** element, each of which is an object contains a **completed** property that equals to `true`. Map method only returns an array of **completed** property which can be anything really (not necessarily the boolean value `true`), NOT **todo** element.